// app/types.ts
export type Agency = {
  name?: string;
  state?: string;
  state_code?: string;
  type?: string;
  population?: string;
  website?: string;
  county?: string;
};

export type Contact = {
  first_name?: string;
  last_name?: string;
  title?: string;
  email?: string;
  phone?: string;
  department?: string;
  email_type?: string;
};